ref: move shmem struct to shmem header
